observable module is renamed to bindable
