+++ pyviews/core/tests/rendering_tests.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock, call
 from pytest import fixture, mark
 
 from pyviews.core.rendering import Node
-from pyviews.core.observable import InheritedDict
+from pyviews.core.bindable import InheritedDict
 from pyviews.core import XmlNode
